requirements.txt
# Japanese Image Translator Dependencies

# Google Cloud Vision API for OCR
google-cloud-vision>=3.4.0
google-api-core>=2.11.0

# OpenAI API for translation
openai>=1.0.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Clustering for text grouping
scikit-learn>=1.3.0

# Environment variables
python-dotenv>=1.0.0